requirements.txt
altgraph==0.17
click==7.1.1
dash==1.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
defusedxml==0.6.0
Flask==1.1.2
Flask-Compress==1.4.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
odfpy==1.4.1
opencv-python==4.2.0.34
plotly==4.6.0
PyInstaller==3.6
PyMuPDF==1.16.16
retrying==1.3.3
six==1.14.0
Werkzeug==1.0.1
